# String optimization