--- a/examples/tui/demoapplication.cpp
+++ b/examples/tui/demoapplication.cpp
@@ -1,6 +1,7 @@
 #include "demoapplication.h"
 
 #include <iostream>
+#include <iomanip>
 #include <chrono>
 #include <thread>