… .github/workflows/examples.yml
… CMakeLists.txt
… examples/gui/CMakeLists.txt
… include/blevalueparser/batteryhealthstatus.h
… include/blevalueparser/batteryinformation.h
… include/blevalueparser/batterylevelstatus.h
… include/blevalueparser/batterytimestatus.h
… include/blevalueparser/bodycompositionfeature.h
… include/blevalueparser/bodycompositionmeasurement.h
… include/blevalueparser/bodysensorlocation.h
… include/blevalueparser/currenttime.h
… include/blevalueparser/dateutc.h
… include/blevalueparser/daydatetime.h
… include/blevalueparser/dayofweek.h
… include/blevalueparser/dstoffset.h
… include/blevalueparser/estimatedservicedate.h
… include/blevalueparser/exacttime256.h
… include/blevalueparser/heartratecontrolpoint.h
… include/blevalueparser/heartratemeasurement.h
… include/blevalueparser/hexstring.h
… include/blevalueparser/localtimeinformation.h
… include/blevalueparser/newalert.h
… include/blevalueparser/pnpid.h
… include/blevalueparser/referencetimeinformation.h
… include/blevalueparser/supportedalertcategorybase.h
… include/blevalueparser/textstring.h
… include/blevalueparser/timeaccuracy.h
… include/blevalueparser/timesource.h
… include/blevalueparser/timezone.h
… include/blevalueparser/unreadalertstatus.h