String optimization

- changed: stringstreams in internal code are replaced with string's `append()` and formatting by `fmt` library for performance purpose
- fixed: magic numbers in TimeAccuracy
- changed: added space beetwen values and full name of units ('123 days' vs '12.3kg')
- changed: hacky architecture fix for prebuilt Qt 5 on Apple Silicon replaced with cmake error with description because hack doesn't work in some cases and looks ugly; QT5_ARCH_ERROR=OFF can be used to bypass error for custom Qt 5 build

Benchmark results:

| Name                            | Before | After  | Diff(ns) | Diff(%) |
|---------------------------------|--------|--------|----------|---------|
| AlertCategoryIDBitMask          |  542ns |  354ns |    188ns |  34.69% |
| AlertCategoryID                 |  107ns | 34.5ns |   72.5ns |  67.76% |
| AlertNotificationControlPoint   |  276ns |  167ns |    109ns |  39.49% |
| BatteryCriticalStatus           |  221ns |  108ns |    113ns |  51.13% |
| BatteryEnergyStatus             | 1148ns |  389ns |    759ns |  66.11% |
| BatteryHealthInformation        |  885ns |  274ns |    611ns |  69.04% |
| BatteryHealthStatus             |  941ns |  344ns |    597ns |  63.44% |
| BatteryInformation              | 3988ns | 1851ns |   2137ns |  53.59% |
| BatteryLevelStatus              | 1041ns |  515ns |    526ns |  50.53% |
| BatteryLevel                    |  171ns | 51.7ns |  119.3ns |  69.77% |
| BatteryTimeStatus               |  505ns |  238ns |    267ns |  52.87% |
| BodyCompositionFeature          |  745ns |  416ns |    329ns |  44.16% |
| BodyCompositionMeasurementMIBFS | 1256ns |  571ns |    685ns |  54.54% |
| BodyCompositionMeasurement      | 3237ns | 1697ns |   1540ns |  47.57% |
| BodySensorLocation              |  106ns | 35.6ns |   70.4ns |  66.42% |
| CurrentTime                     | 1576ns |  484ns |   1092ns |  69.29% |
| DateTime                        |  564ns |  236ns |    328ns |  58.16% |
| DateUTC                         |  843ns |  553ns |    290ns |  34.40% |
| DayDateTime                     |  923ns |  318ns |    605ns |  65.55% |
| DayOfWeek                       |  104ns | 33.4ns |   70.6ns |  67.88% |
| DSTOffset                       |  109ns | 37.2ns |   71.8ns |  65.87% |
| EstimatedServiceDate            | 1103ns |  555ns |    548ns |  49.68% |
| ExactTime256                    | 1202ns |  394ns |    808ns |  67.22% |
| HeartRateControlPoint           |  109ns | 36.6ns |   72.4ns |  66.42% |
| HeartRateMeasurement            | 1043ns |  517ns |    526ns |  50.43% |
| HexString/1                     |  143ns | 49.5ns |   93.5ns |  65.38% |
| HexString/8                     |  248ns |   98ns |    150ns |  60.48% |
| HexString/64                    |  537ns |  195ns |    342ns |  63.69% |
| HexString/512                   | 1920ns |  975ns |    945ns |  49.22% |
| LocalTimeInformation            |  668ns |  155ns |    513ns |  76.80% |
| NewAlert                        |  374ns |  199ns |    175ns |  46.79% |
| PnPID                           |  662ns |  363ns |    299ns |  45.17% |
| ReferenceTimeInformation        |  798ns |  247ns |    551ns |  69.05% |
| SupportedNewAlertCategory       |  893ns |  391ns |    502ns |  56.22% |
| SupportedUnreadAlertCategory    |  890ns |  391ns |    499ns |  56.07% |
| TextString/1                    |   99ns | 25.4ns |   73.6ns |  74.34% |
| TextString/8                    |  107ns | 25.8ns |   81.2ns |  75.89% |
| TextString/64                   |  233ns | 64.9ns |  168.1ns |  72.15% |
| TextString/512                  |  600ns |  137ns |    463ns |  77.17% |
| TimeAccuracy                    |  178ns | 50.2ns |  127.8ns |  71.80% |
| TimeSource                      |  110ns | 37.6ns |   72.4ns |  65.82% |
| TimeZone                        |  154ns | 33.9ns |  120.1ns |  77.99% |
| UnreadAlertStatus               |  271ns |  123ns |    148ns |  54.61% |
| UserIndex                       |  154ns | 32.9ns |  121.1ns |  78.64% |